fix fetch mock for stories
