+++ dotcom-rendering/.storybook/preview.ts
@@ -9,7 +9,6 @@ import { resets } from '@guardian/source/foundations';
 
 import { Lazy } from '../src/components/Lazy';
 import { Picture } from '../src/components/Picture';
-import { mockRESTCalls } from '../src/lib/mockRESTCalls';
 import { setABTests } from '../src/lib/useAB';
 import { ConfigContextDecorator } from './decorators/configContextDecorator';
 import { Preview } from '@storybook/react';
@@ -24,8 +23,6 @@ Picture.disableLazyLoading = isChromatic();
 
 // Fix the date to prevent false negatives
 MockDate.set('Sat Jan 1 2022 12:00:00 GMT+0000 (Greenwich Mean Time)');
-
-mockRESTCalls();
 
 setABTests({
 	api: new AB({
