+++ dotcom-rendering/src/components/ManyNewsletterSignUp.stories.tsx
@@ -1,10 +1,9 @@
-import { mockRESTCalls } from '../lib/mockRESTCalls';
+import { mockedFetch } from '../lib/mockRESTCalls';
 import { GroupedNewslettersList } from './GroupedNewsletterList';
 import { ManyNewsletterSignUp } from './ManyNewsletterSignUp.importable';
 
-mockRESTCalls();
-
 export const Default = () => {
+	global.fetch = mockedFetch;
 	return (
 		<>
 			<GroupedNewslettersList
