+++ dotcom-rendering/src/layouts/DecideLayout.stories.tsx
@@ -17,12 +17,10 @@ import { Video as VideoStandardNewsFixture } from '../../fixtures/generated/dcr-
 import { embedIframe } from '../client/embedIframe';
 import { type ArticleFormat, decideFormat } from '../lib/articleFormat';
 import { getCurrentPillar } from '../lib/layoutHelpers';
-import { mockRESTCalls } from '../lib/mockRESTCalls';
 import { extractNAV } from '../model/extract-nav';
 import type { ArticleDeprecated } from '../types/article';
 import { DecideLayout, type Props as DecideLayoutProps } from './DecideLayout';
-
-mockRESTCalls();
+import { mockedFetch } from '../lib/mockRESTCalls';
 
 export type HydratedLayoutDecoratorArgs = {
 	colourScheme?: 'light' | 'dark';
@@ -39,6 +37,7 @@ export type HydratedLayoutDecoratorArgs = {
 const HydratedLayout: Decorator<
 	DecideLayoutProps & HydratedLayoutDecoratorArgs
 > = (Story, context) => {
+	global.fetch = mockedFetch;
 	const { article } = context.args;
 	const format: ArticleFormat = decideFormat(article.format);
 	const colourScheme =
