+++ dotcom-rendering/src/lib/mockRESTCalls.ts
@@ -1,4 +1,3 @@
-import fetchMock from 'fetch-mock';
 import { matchReport } from '../../fixtures/generated/match-report';
 import { series } from '../../fixtures/generated/series';
 import { storyPackage } from '../../fixtures/generated/story-package';
@@ -15,6 +14,240 @@ import { noTopPicks } from '../../fixtures/manual/noTopPicks';
 import { related } from '../../fixtures/manual/related';
 import { shortDiscussion } from '../../fixtures/manual/short-discussion';
 import { topPicks } from '../../fixtures/manual/topPicks';
+
+const createMockResponse = (status: number, body?: any): Promise<Response> => {
+	return Promise.resolve(
+		new Response(JSON.stringify(body), { status: status }),
+	);
+};
+
+let lastFetchCall: [url: string, requestInit?: RequestInit | undefined] | null =
+	null;
+
+export const mockedFetch = (
+	input: RequestInfo | URL,
+	requestInit?: RequestInit,
+) => {
+	const url = input.toString();
+	lastFetchCall = [url, requestInit];
+
+	console.log(`fetched url: ${url}`);
+	console.log(`fetched method: ${requestInit?.method}`);
+	switch (true) {
+		case /logs\.(code.dev-)?guardianapis\.com\/log/.test(url) &&
+			requestInit?.method === 'POST':
+			return createMockResponse(200);
+		case /.*api.nextgen.guardianapps.co.uk\/most-read-geo.*/.test(url):
+			return createMockResponse(200, mostReadGeo);
+		// Comment count
+		case /.*api.nextgen.guardianapps.co.uk\/discussion\/comment-counts.*/.test(
+			url,
+		):
+			return createMockResponse(200, {
+				counts: [
+					{
+						id: '/p/4k83z',
+						count: 432,
+					},
+				],
+			});
+		case /\/discussion\/comments\/count.*/.test(url):
+			return createMockResponse(200, {
+				counts: [
+					{
+						id: '/p/4k83z',
+						count: 432,
+					},
+				],
+			});
+		case /^https:\/\/discussion\.theguardian\.com\/discussion-api\/getCommentCounts\?/.test(
+			url,
+		):
+			return createMockResponse(200, {
+				'/p/d8ex5': 496,
+				'/p/zemg8': 11_000,
+			});
+		case /^https:\/\/discussion\.theguardian\.com\/discussion-api\/getCommentCounts\?short-urls=\/p\/d8ex5/.test(
+			url,
+		):
+			return createMockResponse(200, {
+				'/p/d8ex5': 496,
+			});
+		// Most read by category
+		case /.*api.nextgen.guardianapps.co.uk\/most-read.*/.test(url):
+			return createMockResponse(200, mostRead);
+		// Related
+		case /.*api.nextgen.guardianapps.co.uk\/related.*/.test(url):
+			return createMockResponse(200, related);
+		// Popular in tag
+		case /.*api.nextgen.guardianapps.co.uk\/popular-in-tag.*/.test(url):
+			return createMockResponse(200, related);
+		// Series
+		case /.*api.nextgen.guardianapps.co.uk\/series.*/.test(url):
+			return createMockResponse(200, series);
+		// Story package
+		case /.*api.nextgen.guardianapps.co.uk\/story-package.*/.test(url):
+			return createMockResponse(200, storyPackage);
+		// Rich link
+		case /.*api.nextgen.guardianapps.co.uk\/embed\/card.*/.test(url):
+			return createMockResponse(200, richLinkCard);
+		// Get 'short' discussion - top comments
+		case /.*discussion.theguardian.com\/discussion-api\/discussion\/p\/4v8kk\/topcomments/.test(
+			url,
+		):
+			return createMockResponse(200, discussionNoTopComments);
+		// Get 'short' discussion
+		case /.*discussion.theguardian.com\/discussion-api\/discussion\/p\/4v8kk/.test(
+			url,
+		):
+			return createMockResponse(200, shortDiscussion);
+		// Get country code
+		case /.*api.nextgen.guardianapps.co.uk\/geolocation.*/.test(url):
+			return createMockResponse(200, { country: 'GB' });
+		// Match report data
+		case /.*api.nextgen.guardianapps.co.uk\/football\/api.*/.test(url):
+			return createMockResponse(200, matchReport);
+		// Get user discussion api (used for myAccount dropdown)
+		case /discussionApiUrl\/profile\/me\?strict_sanctions_check=false/.test(
+			url,
+		):
+			return createMockResponse(200, {
+				status: 'ok',
+				userProfile: {
+					userId: '123',
+					displayName: 'Guardian User',
+					webUrl: 'https://profile.test-theguardian.com/user/id/123',
+					apiUrl: 'http://discussion.test-guardianapis.com/discussion-api/profile/123',
+					avatar: 'https://avatar.test-guimcode.co.uk/user/123',
+					secureAvatarUrl:
+						'https://avatar.test-guimcode.co.uk/user/123',
+					badge: [],
+					privateFields: {
+						canPostComment: true,
+						isPremoderated: false,
+						hasCommented: false,
+					},
+				},
+			});
+		// Get discussion 39f5z
+		case /.*\/discussion.theguardian.com\/discussion-api\/discussion\/p\/39f5z\?.*/.test(
+			url,
+		):
+			return createMockResponse(200, discussion);
+		case /.*\/discussion\/p\/39f5z\/topcomments.*/.test(url):
+			return createMockResponse(200, topPicks);
+		// Get discussion 39f5x
+		case /.*\/discussion.theguardian.com\/discussion-api\/discussion\/p\/39f5x\?.*/.test(
+			url,
+		):
+			return createMockResponse(200, discussionWithNoComments);
+		case /.*\/discussion\/p\/39f5x\/topcomments.*/.test(url):
+			return createMockResponse(200, noTopPicks);
+		// Get discussion 39f5a
+		case /.*\/discussion.theguardian.com\/discussion-api\/discussion\/p\/39f5a\?.*/.test(
+			url,
+		):
+			return createMockResponse(200, discussionWithTwoComments);
+		case /.*\/discussion\/p\/39f5a\/topcomments.*/.test(url):
+			return createMockResponse(200, noTopPicks);
+		// Get discussion 32255
+		case /.*\/discussion.theguardian.com\/discussion-api\/discussion\/p\/32255\?.*/.test(
+			url,
+		):
+			return createMockResponse(200, legacyDiscussionWithoutThreading);
+		case /.*\/discussion\/p\/32255\/topcomments.*/.test(url):
+			return createMockResponse(200, noTopPicks);
+		// Get discussion abc123
+		case /.*\/discussion.theguardian.com\/discussion-api\/discussion\/p\/abc123\?.*/.test(
+			url,
+		):
+			return createMockResponse(200, discussion);
+		case /.*\/discussion\/p\/abc123\/topcomments.*/.test(url):
+			return createMockResponse(200, noTopPicks);
+		// Get more replies
+		case /.*discussion.theguardian.com\/discussion-api\/comment\/.*/.test(
+			url,
+		):
+			return createMockResponse(200, {
+				status: 'ok',
+				comment,
+			});
+		// Recommend
+		case /.*discussion.theguardian.com\/discussion-api\/comment\/.*\/recommend/.test(
+			url,
+		) && requestInit?.method === 'POST':
+			return createMockResponse(200, {
+				status: 'ok',
+			});
+		// Abuse form
+		case /.*discussion.theguardian.com\/discussion-api\/comment\/.*\/reportAbuse/.test(
+			url,
+		) && requestInit?.method === 'POST':
+			return createMockResponse(200, {
+				status: 'ok',
+			});
+		// Login redirect
+		case /.*profile\.theguardian\.com\/signin\?INTCMP=DOTCOM_NEWHEADER_SIGNIN/.test(
+			url,
+		):
+			return createMockResponse(200);
+		case /.*profile\.theguardian\.com\/actions\/signInSecondStepCurrent/.test(
+			url,
+		) && requestInit?.method === 'POST':
+			return createMockResponse(200);
+		//https://discussion.theguardian.com/discussion-api/discussion/p/g8g7v/comment?api-key=dotcom-rendering
+		// This returns a mocked error for a story 97d6eab4a98917f63bc96a7ac64f7ca7
+		case /.*discussion.theguardian.com\/discussion-api\/discussion\/p\/g8g7v\/.*/.test(
+			url,
+		) && requestInit?.method === 'POST':
+			return createMockResponse(400, {
+				status: 'error',
+				message: 'API: Username Missing',
+				errorCode: 'USERNAME_MISSING',
+			});
+		// Get discussion
+		case /.*discussion.theguardian.com\/discussion-api\/discussion\/.*/.test(
+			url,
+		):
+			return createMockResponse(200, discussion);
+		case /.*contributions\.(code.dev-)?guardianapis\.com\/header/.test(
+			url,
+		) && requestInit?.method === 'POST':
+			return createMockResponse(200, contributionsHeaderResponse);
+		// Get Ophan
+		case /.*ophan.theguardian.com\/img\/.*/.test(url):
+			return createMockResponse(200);
+
+		// Return an error response if the request body includes the
+		// phrase 'example.com', otherwise, return a success response.
+		// For use on stories and tests involving posts to the '/email/many'
+		// or '/email'
+		// api endpoint eg:
+		// dotcom-rendering/src/components/ManyNewsletterSignUp.stories.tsx
+		case /.*api.nextgen.guardianapps.co.uk\/email[/many]{0,1}.*/.test(
+			url,
+		) && requestInit?.method === 'POST':
+			const decodedBody = decodeURIComponent(
+				requestInit?.body?.toString() ?? '',
+			);
+			return decodedBody.includes('example.com')
+				? createMockResponse(500)
+				: createMockResponse(200);
+		default:
+			console.log('could not find url');
+			return Promise.resolve(
+				new Response(JSON.stringify({ error: 'Not Found' }), {
+					status: 404,
+				}),
+			);
+	}
+};
+
+export const getLastFetchCall = () => lastFetchCall;
+
+export const resetLastFetchCall = () => {
+	lastFetchCall = null;
+};
 
 const richLinkCard = {
 	tags: [
@@ -124,395 +357,4 @@ const richLinkCard = {
 		theme: 'CulturePillar',
 		display: 'StandardDisplay',
 	},
-};
-
-export const mockRESTCalls = (): typeof fetchMock => {
-	return (
-		fetchMock
-			.restore()
-			// Intercept Logs
-			.post(/logs\.(code.dev-)?guardianapis\.com\/log/, {
-				status: 204,
-			})
-			// Most read by Geo
-			.get(
-				/.*api.nextgen.guardianapps.co.uk\/most-read-geo.*/,
-				{
-					status: 200,
-					body: mostReadGeo,
-				},
-				{ overwriteRoutes: false },
-			)
-			// Comment count
-			.get(
-				/.*api.nextgen.guardianapps.co.uk\/discussion\/comment-counts.*/,
-				{
-					status: 200,
-					body: {
-						counts: [
-							{
-								id: '/p/4k83z',
-								count: 432,
-							},
-						],
-					},
-				},
-				{ overwriteRoutes: false },
-			)
-			.get(
-				/\/discussion\/comments\/count.*/,
-				{
-					status: 200,
-					body: {
-						counts: [
-							{
-								id: '/p/4k83z',
-								count: 432,
-							},
-						],
-					},
-				},
-				{ overwriteRoutes: false },
-			)
-			.get(
-				/^https:\/\/discussion\.theguardian\.com\/discussion-api\/getCommentCounts\?/,
-				{
-					status: 200,
-					body: {
-						'/p/d8ex5': 496,
-						'/p/zemg8': 11_000,
-					},
-				},
-				{ overwriteRoutes: false },
-			)
-			.get(
-				/^https:\/\/discussion\.theguardian\.com\/discussion-api\/getCommentCounts\?short-urls=\/p\/d8ex5/,
-				{
-					status: 200,
-					body: {
-						'/p/d8ex5': 496,
-					},
-				},
-				{ overwriteRoutes: false },
-			)
-			// Most read by category
-			.get(
-				/.*api.nextgen.guardianapps.co.uk\/most-read.*/,
-				{
-					status: 200,
-					body: mostRead,
-				},
-				{ overwriteRoutes: false },
-			)
-			// Related
-			.get(
-				/.*api.nextgen.guardianapps.co.uk\/related.*/,
-				{
-					status: 200,
-					body: related,
-				},
-				{ overwriteRoutes: false },
-			)
-			// Popular in tag
-			.get(
-				/.*api.nextgen.guardianapps.co.uk\/popular-in-tag.*/,
-				{
-					status: 200,
-					body: related,
-				},
-				{ overwriteRoutes: false },
-			)
-			// Series
-			.get(
-				/.*api.nextgen.guardianapps.co.uk\/series.*/,
-				{
-					status: 200,
-					body: series,
-				},
-				{ overwriteRoutes: false },
-			)
-			// Story package
-			.get(
-				/.*api.nextgen.guardianapps.co.uk\/story-package.*/,
-				{
-					status: 200,
-					body: storyPackage,
-				},
-				{ overwriteRoutes: false },
-			)
-			// Rich link
-			.get(
-				/.*api.nextgen.guardianapps.co.uk\/embed\/card.*/,
-				{
-					status: 200,
-					body: richLinkCard,
-				},
-				{ overwriteRoutes: false },
-			)
-			// Get 'short' discussion - top comments
-			.get(
-				/.*discussion.theguardian.com\/discussion-api\/discussion\/p\/4v8kk\/topcomments/,
-				{
-					status: 200,
-					body: discussionNoTopComments,
-				},
-				{ overwriteRoutes: false },
-			)
-			// Get 'short' discussion
-			.get(
-				/.*discussion.theguardian.com\/discussion-api\/discussion\/p\/4v8kk/,
-				{
-					status: 200,
-					body: shortDiscussion,
-				},
-				{ overwriteRoutes: false },
-			)
-			// Get country code
-			.get(
-				/.*api.nextgen.guardianapps.co.uk\/geolocation.*/,
-				{
-					status: 200,
-					body: { country: 'GB' },
-				},
-				{ overwriteRoutes: false },
-			)
-			// Match report data
-			.get(
-				/.*api.nextgen.guardianapps.co.uk\/football\/api.*/,
-				{
-					status: 200,
-					body: matchReport,
-				},
-				{ overwriteRoutes: false },
-			)
-
-			// Get user discussion api (used for myAccount dropdown)
-			.get(
-				/discussionApiUrl\/profile\/me\?strict_sanctions_check=false/,
-				{
-					status: 200,
-					body: {
-						status: 'ok',
-						userProfile: {
-							userId: '123',
-							displayName: 'Guardian User',
-							webUrl: 'https://profile.test-theguardian.com/user/id/123',
-							apiUrl: 'http://discussion.test-guardianapis.com/discussion-api/profile/123',
-							avatar: 'https://avatar.test-guimcode.co.uk/user/123',
-							secureAvatarUrl:
-								'https://avatar.test-guimcode.co.uk/user/123',
-							badge: [],
-							privateFields: {
-								canPostComment: true,
-								isPremoderated: false,
-								hasCommented: false,
-							},
-						},
-					},
-				},
-				{ overwriteRoutes: false },
-			)
-			// Get discussion 39f5z
-			.get(
-				/.*\/discussion.theguardian.com\/discussion-api\/discussion\/p\/39f5z\?.*/,
-				{
-					status: 200,
-					body: discussion,
-				},
-				{ overwriteRoutes: false },
-			)
-			.get(
-				/.*\/discussion\/p\/39f5z\/topcomments.*/,
-				{
-					status: 200,
-					body: topPicks,
-				},
-				{ overwriteRoutes: false },
-			)
-
-			// Get discussion 39f5x
-			.get(
-				/.*\/discussion.theguardian.com\/discussion-api\/discussion\/p\/39f5x\?.*/,
-				{
-					status: 200,
-					body: discussionWithNoComments,
-				},
-				{ overwriteRoutes: false },
-			)
-			.get(
-				/.*\/discussion\/p\/39f5x\/topcomments.*/,
-				{
-					status: 200,
-					body: noTopPicks,
-				},
-				{ overwriteRoutes: false },
-			)
-
-			// Get discussion 39f5a
-			.get(
-				/.*\/discussion.theguardian.com\/discussion-api\/discussion\/p\/39f5a\?.*/,
-				{
-					status: 200,
-					body: discussionWithTwoComments,
-				},
-				{ overwriteRoutes: false },
-			)
-			.get(
-				/.*\/discussion\/p\/39f5a\/topcomments.*/,
-				{
-					status: 200,
-					body: noTopPicks,
-				},
-				{ overwriteRoutes: false },
-			)
-
-			// Get discussion 32255
-			.get(
-				/.*\/discussion.theguardian.com\/discussion-api\/discussion\/p\/32255\?.*/,
-				{
-					status: 200,
-					body: legacyDiscussionWithoutThreading,
-				},
-				{ overwriteRoutes: false },
-			)
-			.get(
-				/.*\/discussion\/p\/32255\/topcomments.*/,
-				{
-					status: 200,
-					body: noTopPicks,
-				},
-				{ overwriteRoutes: false },
-			)
-
-			// Get discussion abc123
-			.get(
-				/.*\/discussion.theguardian.com\/discussion-api\/discussion\/p\/abc123\?.*/,
-				{
-					status: 200,
-					body: discussion,
-				},
-				{ overwriteRoutes: false },
-			)
-			.get(
-				/.*\/discussion\/p\/abc123\/topcomments.*/,
-				{
-					status: 200,
-					body: noTopPicks,
-				},
-				{ overwriteRoutes: false },
-			)
-
-			// Get more replies
-			.get(
-				/.*discussion.theguardian.com\/discussion-api\/comment\/.*/,
-				{
-					status: 200,
-					body: {
-						status: 'ok',
-						comment,
-					},
-				},
-				{ overwriteRoutes: false },
-			)
-
-			// Recommend
-			.post(
-				/.*discussion.theguardian.com\/discussion-api\/comment\/.*\/recommend/,
-				{
-					status: 200,
-					body: {
-						status: 'ok',
-					},
-				},
-				{ overwriteRoutes: false },
-			)
-
-			// Abuse form
-			.post(
-				/.*discussion.theguardian.com\/discussion-api\/comment\/.*\/reportAbuse/,
-				{
-					status: 200,
-					body: {
-						status: 'ok',
-					},
-				},
-				{ overwriteRoutes: false },
-			)
-
-			// Login redirect
-			.get(
-				/.*profile\.theguardian\.com\/signin\?INTCMP=DOTCOM_NEWHEADER_SIGNIN/,
-				{
-					status: 200,
-				},
-				{ overwriteRoutes: false },
-			)
-			.post(
-				/.*profile\.theguardian\.com\/actions\/signInSecondStepCurrent/,
-				{
-					status: 200,
-				},
-			)
-
-			//https://discussion.theguardian.com/discussion-api/discussion/p/g8g7v/comment?api-key=dotcom-rendering
-			// This returns a mocked error for a story 97d6eab4a98917f63bc96a7ac64f7ca7
-			.post(
-				/.*discussion.theguardian.com\/discussion-api\/discussion\/p\/g8g7v\/.*/,
-				{
-					status: 400,
-					body: {
-						status: 'error',
-						message: 'API: Username Missing',
-						errorCode: 'USERNAME_MISSING',
-					},
-				},
-				{ overwriteRoutes: false },
-			)
-			// Get discussion
-			.get(
-				/.*discussion.theguardian.com\/discussion-api\/discussion\/.*/,
-				{
-					status: 200,
-					body: discussion,
-				},
-				{ overwriteRoutes: false },
-			)
-			.post(
-				/.*contributions\.(code.dev-)?guardianapis\.com\/header/,
-				{
-					status: 200,
-					body: contributionsHeaderResponse,
-				},
-				{ overwriteRoutes: false },
-			)
-			// Get Ophan
-			.get(
-				/.*ophan.theguardian.com\/img\/.*/,
-				{ status: 200 },
-				{ overwriteRoutes: false },
-			)
-
-			// Return an error response if the request body includes the
-			// phrase 'example.com', otherwise, return a success response.
-			// For use on stories and tests involving posts to the '/email/many'
-			// or '/email'
-			// api endpoint eg:
-			// dotcom-rendering/src/components/ManyNewsletterSignUp.stories.tsx
-			.post(
-				/.*api.nextgen.guardianapps.co.uk\/email[/many]{0,1}.*/,
-				(url, mockRequest) => {
-					const decodedBody = decodeURIComponent(
-						mockRequest.body?.toString() ?? '',
-					);
-
-					return decodedBody.includes('example.com')
-						? { status: 500 }
-						: { status: 200 };
-				},
-				{ overwriteRoutes: false },
-			)
-
-			/** @see https://github.com/wheresrhys/fetch-mock/issues/618 */
-			.spy('end:.hot-update.json')
-	);
 };
